pyproject.toml
[tool.poetry]
name = "basic-name"
version = "0.1.0"
description = ""
authors = [
    "GodVov4 <volume245@gmail.com>",
    "DSofiya <dsonya94@gmail.com>",
    "fls0 <sadurskyi.m@gmail.com>",
    "GennadiiBukh <gennadii.bukh@gmail.com>"
]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "0.108.0"
uvicorn = {extras = ["standard"], version = "0.25.0"}
sqlalchemy = "2.0.25"
python-multipart = "0.0.6"
python-jose = {extras = ["cryptography"], version = "3.3.0"}
passlib = {extras = ["bcrypt"], version = "1.7.4"}
cloudinary = "1.38.0"
alembic = "1.13.1"
asyncpg = "0.29.0"
pydantic = {extras = ["email"], version = "2.5.3"}
fastapi-mail = "1.4.1"
redis = "5.0.1"
fastapi-limiter = "0.1.6"
libgravatar = "1.0.4"
qrcode = "7.4.2"
pillow = "10.2.0"


[tool.poetry.group.test.dependencies]
pytest = "7.4.4"
httpx = "^0.26.0"


[tool.poetry.group.dev.dependencies]
sphinx = "7.2.6"
sphinx-press-theme = "0.8.0"
myst-parser = "2.0.0"
sphinx-autodoc2 = "0.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
